+++ src/resolvers/answer.ts
@@ -53,8 +53,8 @@ export const resolvers: Resolvers = {
     },
 
     // return the answer for the given versionedQuestionId
-    answerByVersionedQuestionId: async (_, { planId, versionedQuestionId }, context: MyContext): Promise<Answer> => {
-      const reference = 'planSectionAnswers resolver';
+    answerByVersionedQuestionId: async (_, { planId, versionedQuestionId, versionedCustomQuestionId }, context: MyContext): Promise<Answer> => {
+      const reference = 'answerByVersionedQuestionId resolver';
       try {
         if (isAuthorized(context.token)) {
           const plan = await Plan.findById(reference, context, planId);
@@ -63,8 +63,14 @@ export const resolvers: Resolvers = {
           }
           const project = await Project.findById(reference, context, plan.projectId);
           if (await hasPermissionOnProject(context, project, ProjectCollaboratorAccessLevel.COMMENT)) {
-            const temp = await Answer.findByPlanIdAndVersionedQuestionId(reference, context, planId, versionedQuestionId);
-            return temp;
+            if (versionedCustomQuestionId) {
+              return await Answer.findByPlanIdAndVersionedCustomQuestionId(
+                reference, context, planId, versionedCustomQuestionId
+              );
+            }
+            return await Answer.findByPlanIdAndVersionedQuestionId(
+              reference, context, planId, versionedQuestionId
+            );
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -103,7 +109,7 @@ export const resolvers: Resolvers = {
 
   Mutation: {
     // Create a new answer
-    addAnswer: async (_, { planId, versionedSectionId, versionedQuestionId, json }, context: MyContext): Promise<Answer> => {
+    addAnswer: async (_, { planId, versionedSectionId, versionedQuestionId, versionedCustomSectionId, versionedCustomQuestionId, json }, context: MyContext): Promise<Answer> => {
       const reference = 'addAnswer resolver';
       try {
         if (isAuthorized(context.token)) {
@@ -114,7 +120,7 @@ export const resolvers: Resolvers = {
 
           const project = await Project.findById(reference, context, plan.projectId);
           if (await hasPermissionOnProject(context, project)) {
-            const answer = new Answer({ planId, versionedSectionId, versionedQuestionId, json });
+            const answer = new Answer({ planId, versionedSectionId, versionedQuestionId, versionedCustomSectionId, versionedCustomQuestionId, json });
             const newAnswer = await answer.create(context);
             if (newAnswer && !newAnswer.hasErrors()) {
               // Update the maDMP version of the Plan